Moved Iterator serializers into single factory class
