+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/serialization/kryo/LinkedHashMapEntrySerializer.kt
@@ -4,7 +4,6 @@ import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.Serializer
 import com.esotericsoftware.kryo.io.Input
 import com.esotericsoftware.kryo.io.Output
-import java.util.LinkedList
 
 /**
  * The [LinkedHashMap] and [LinkedHashSet] have a problem with the default Quasar/Kryo serialisation
@@ -35,25 +34,3 @@ object LinkedHashMapEntrySerializer : Serializer<Map.Entry<*, *>>() {
         return constructor.newInstance(0, key, value, null) as Map.Entry<*, *>
     }
 }
-
-/**
- * Also, add a [ListIterator] serializer to avoid more linked list issues.
-*/
-object LinkedListItrSerializer : Serializer<ListIterator<Any>>() {
-    // Create a dummy list so that we can get the ListItr from it
-    // The element type of the list doesn't matter.  The iterator is all we want
-    private val outerListField = LinkedList<Long>(listOf(1)).listIterator()::class.java.getDeclaredField("this$0").apply { isAccessible = true }
-
-    override fun write(kryo: Kryo, output: Output, obj: ListIterator<Any>) {
-        kryo.writeClassAndObject(output, outerListField.get(obj))
-        output.writeInt(obj.nextIndex())
-    }
-
-    override fun read(kryo: Kryo, input: Input, type: Class<out ListIterator<Any>>): ListIterator<Any> {
-        val list = kryo.readClassAndObject(input) as LinkedList<*>
-        val index = input.readInt()
-        return list.listIterator(index)
-    }
-}
-
-
